+++ lib/player.ex
@@ -1,0 +1,3 @@
+defprotocol TicTacToe.Player do
+  def choose_move(player, board)
+end
