+++ lib/player_minimax.ex
@@ -5,7 +5,7 @@ defmodule PlayerMinimax do
     do: %PlayerMinimax{player: player_mark, opponent: opponent_mark}
 end
 
-defimpl Player, for: PlayerMinimax do
+defimpl TicTacToe.Player, for: PlayerMinimax do
   def choose_move(%PlayerMinimax{player: player, opponent: opponent}, board) do
     Minimax.new(board, player, opponent)
     |> Minimax.best_position()
